+++ packages/cli/src/deploy/deploy.ts
@@ -14,7 +14,7 @@ import { debug } from "./debug";
 import { resourceLabel } from "./resourceLabel";
 import { uniqueBy } from "@latticexyz/common/utils";
 import { ensureContractsDeployed } from "./ensureContractsDeployed";
-import { coreModuleBytecode, worldFactoryBytecode } from "./ensureWorldFactory";
+import { coreModuleBytecode, worldFactoryBytecode, worldFactoryContracts } from "./ensureWorldFactory";
 
 type DeployOptions<configInput extends ConfigInput> = {
   client: Client<Transport, Chain | undefined, Account>;
@@ -45,6 +45,7 @@ export async function deploy<configInput extends ConfigInput>({
     contracts: [
       ...uniqueBy(libraries, (library) => getAddress(library.address)).map((library) => ({
         bytecode: library.bytecode,
+        deployedBytecodeSize: library.deployedBytecodeSize,
         label: `${library.fullyQualifiedName} library`,
       })),
     ],
@@ -54,14 +55,15 @@ export async function deploy<configInput extends ConfigInput>({
   await ensureContractsDeployed({
     client,
     contracts: [
-      { bytecode: coreModuleBytecode, label: "core module" },
-      { bytecode: worldFactoryBytecode, label: "world factory" },
+      ...worldFactoryContracts,
       ...uniqueBy(systems, (system) => getAddress(system.address)).map((system) => ({
         bytecode: system.bytecode,
+        deployedBytecodeSize: system.deployedBytecodeSize,
         label: `${resourceLabel(system)} system`,
       })),
       ...uniqueBy(config.modules, (mod) => getAddress(mod.address)).map((mod) => ({
         bytecode: mod.bytecode,
+        deployedBytecodeSize: mod.deployedBytecodeSize,
         label: `${mod.name} module`,
       })),
     ],