+++ packages/dev-tools/src/router.tsx
@@ -6,6 +6,8 @@ import { SummaryPage } from "./summary/SummaryPage";
 import { ActionsPage } from "./actions/ActionsPage";
 import { ComponentsPage } from "./recs/ComponentsPage";
 import { ComponentData } from "./recs/ComponentData";
+import { TablesPage } from "./zustand/TablesPage";
+import { TableData } from "./zustand/TableData";
 
 export const router = createMemoryRouter(
   createRoutesFromElements(
@@ -13,6 +15,9 @@ export const router = createMemoryRouter(
       <Route index element={<SummaryPage />} />
       <Route path="actions" element={<ActionsPage />} />
       <Route path="events" element={<EventsPage />} />
+      <Route path="tables" element={<TablesPage />}>
+        <Route path=":id" element={<TableData />} />
+      </Route>
       <Route path="components" element={<ComponentsPage />}>
         <Route path=":id" element={<ComponentData />} />
       </Route>